+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Added
 - Added MQTT polling source support 
+- Added MQTT metadata block support
 
 ## [0.19.0] - 2024-05-14
 ### Added